refactor: move venv/controller logic from PythonProvider to PythonPipProvider
PythonProvider now only holds shared utilities (toPurl, handleIgnoredDependencies,
containsIgnorePattern, root component defaults). The venv-based provideStack/
provideComponent, getPythonController, getExecutable, and related methods are
moved to PythonPipProvider — the only consumer.
